+++ tests/test_metadata.py
@@ -113,6 +113,7 @@ expected = {
     "python_files": [
         "gha_utils/__init__.py",
         "gha_utils/__main__.py",
+        "gha_utils/bumpversion.py",
         "gha_utils/changelog.py",
         "gha_utils/cli.py",
         "gha_utils/labels.py",
@@ -123,6 +124,7 @@ expected = {
         "gha_utils/test_plan.py",
         "gha_utils/workflows.py",
         "tests/__init__.py",
+        "tests/test_bumpversion.py",
         "tests/test_changelog.py",
         "tests/test_mailmap.py",
         "tests/test_matrix.py",
@@ -150,6 +152,7 @@ expected = {
         "tests/cli-test-plan.yaml",
     ],
     "toml_files": [
+        ".github/bumpversion.toml",
         ".github/labels.toml",
         "lychee.toml",
         "pyproject.toml",